refactor product test to a hook
